requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
ollama
openai
mysql-connector-python
nest_asyncio
